+++ CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated documentation.
+- SAPI client has been replaced with new Zumokit API client.
 - AuthController and HealthCheckController now uses new Zumokit API client.
+- Excess code cleanup.
 
 ### Removed